requirements.txt
# Python 3.8.16

annotated-types==0.6.0
anyio==4.3.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
blinker==1.5
Brotli==1.0.9
bs4==0.0.1
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.0.1
Click==7.0
clickhouse-driver==0.1.4
dash==2.15.0
dash-bootstrap-components==0.10.7
dash-core-components==2.0.0
dash-extensions==0.0.31
dash-html-components==2.0.0
dash-renderer==1.5.1
dash-table==5.0.0
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
Flask==1.1.2
Flask-Caching==1.9.0
Flask-Compress==1.5.0
Flask-SQLAlchemy==2.4.4
future==0.18.2
gevent==22.10.2
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==2.10
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
lxml==4.9.2
MarkupSafe==1.1.1
more-itertools==9.0.0
multitasking==0.0.9
numpy==1.19.2
openai==1.12.0
openpyxl==3.0.3
pandas==1.1.2
pandas-datareader==0.9.0
plotly==5.19.0
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
PyMySQL==0.9.3
python-dotenv==1.0.0
pytz==2019.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
retrying==1.3.3
sentry-sdk==1.13.0
simplejson==3.17.2
sniffio==1.3.0
soupsieve==2.0.1
SQLAlchemy==1.3.13
sqlalchemy-vertica==0.0.5
sqlalchemy-vertica-python==0.5.9
tenacity==8.2.3
tqdm==4.66.2
tushare==1.2.60
typing_extensions==4.9.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.11
vertica-python==1.2.0
websocket-client==1.4.2
Werkzeug==1.0.0
yfinance==0.1.54
zope.event==4.6
zope.interface==5.5.2
